+++ CHANGELOG.md
@@ -41,6 +41,7 @@
 * [BUGFIX] fix: normalize allowlist headers when building the allowlist map [#6481](https://github.com/grafana/tempo/pull/6481) (@javiermolinar)
 * [BUGFIX] fix: compactor deduped spans metric uses wrong type (gauge instead of counter) [#6558](https://github.com/grafana/tempo/issues/6558) (@bejaratommy)
 * [BUGFIX] metrics-generator: Fix active-series counter underflow in local series limiter when overflow series are deleted [#6568](https://github.com/grafana/tempo/pull/6568) (@carles-grafana)
+* [BUGFIX] livestore: avoid panic on query lag checks during startup/shutdown [#6567](https://github.com/grafana/tempo/pull/6567) (@carles-grafana)
 
 ### 3.0 Cleanup